move consumertest into per-signal modules
