--- consumer/consumertest/doc.go
@@ -1,6 +1,0 @@
-// Copyright The OpenTelemetry Authors
-// SPDX-License-Identifier: Apache-2.0
-
-// Package consumertest defines types and functions used to help test packages
-// implementing the consumer package interfaces.
-package consumertest // import "go.opentelemetry.io/collector/consumer/consumertest"
